fix(infra-monitoring): rename entityConfig to entity.config
