+++ frontend/src/container/InfraMonitoringK8sV2/Base/entity.config.types.ts
